refactor(cache): move the generic formatters into app/src/utils
humanizeSeconds was a general duration formatter buried in the cache view; extract it as app/src/utils/format-duration.ts (formatDuration) with its own test, following the repo's one-formatter-per-file + JSDoc convention. Drop the bespoke formatSize in favour of the existing formatFilesize util (byte labels shift KB -> kB, decimal default). The relative-time helpers (formatAge/Expiry/LastHit) stay in cache-view — they carry the feature's now/label semantics.
